+++ ribs/archives/__init__.py
@@ -14,6 +14,7 @@ inheriting from :class:`~ribs.archives.ArchiveBase`.
     CategoricalArchive
     CVTArchive
     DensityArchive
+    DiscountArchive
     DNSArchive
     GridArchive
     ProximityArchive
@@ -46,6 +47,7 @@ from ribs.archives._categorical_archive import CategoricalArchive
 from ribs.archives._cqd_score import CQDScoreResult, cqd_score
 from ribs.archives._cvt_archive import CVTArchive, k_means_centroids
 from ribs.archives._density_archive import DensityArchive
+from ribs.archives._discount_archive import DiscountArchive
 from ribs.archives._dns_archive import DNSArchive
 from ribs.archives._grid_archive import GridArchive
 from ribs.archives._proximity_archive import ProximityArchive
@@ -56,6 +58,7 @@ __all__ = [
     "CategoricalArchive",
     "CVTArchive",
     "DensityArchive",
+    "DiscountArchive",
     "DNSArchive",
     "GridArchive",
     "ProximityArchive",