+++ tests/emitters/evolution_strategy_emitter_test.py
@@ -9,6 +9,8 @@ from ribs.emitters import EvolutionStrategyEmitter
 RANKER_LIST = ["imp", "2imp", "rd", "2rd", "obj", "2obj"]
 ES_LIST = ["cma_es", "sep_cma_es", "lm_ma_es", "openai_es"]
 
+# pylint: disable=redefined-outer-name
+
 
 @pytest.fixture
 def emitter_fixture(request):