+++ tests/visualize/parallel_axes_plot_test.py
@@ -10,12 +10,16 @@ from ribs.visualize import parallel_axes_plot
 
 # See https://github.com/astral-sh/ruff/issues/10662
 # ruff: noqa: F401, F811
-from .grid_archive_heatmap_test import (
+from .grid_archive_heatmap_test import (  # pylint: disable = unused-import
     grid_archive_2d,
     grid_archive_3d,
     grid_archive_3d_empty,
 )
-from .proximity_archive_plot_test import proximity_archive_2d_obj
+from .proximity_archive_plot_test import (  # pylint: disable = unused-import
+    proximity_archive_2d_obj,
+)
+
+# pylint: disable=redefined-outer-name
 
 
 @image_comparison(baseline_images=["2d"], remove_text=False, extensions=["png"])