+++ tests/visualize/proximity_archive_plot_test.py
@@ -11,6 +11,8 @@ from matplotlib.testing.decorators import image_comparison
 from ribs.archives import ProximityArchive
 from ribs.visualize import proximity_archive_plot
 
+# pylint: disable=redefined-outer-name
+
 #
 # Fixtures
 #